% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_demo_data}
\alias{ems_demo_data}
\title{EMS demo data.}
\format{A data frame with 907694 rows and 22 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
ems_demo_data
}
\description{
EMS data from 2018-01-01 to 2019-03-12.
}
\keyword{datasets}
