% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_demo_parameters}
\alias{ems_demo_parameters}
\title{EMS parameters.}
\format{A data frame with 907694 rows and 22 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
ems_demo_parameters
}
\description{
EMS lookup table for which sites have data for each parameter.
}
\keyword{datasets}
