% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_demo_sites}
\alias{ems_demo_sites}
\title{EMS sites.}
\format{A data frame with 907694 rows and 22 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
ems_demo_sites
}
\description{
EMS lookup table for sites and their location.
}
\keyword{datasets}
