% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_site_parameters}
\alias{ems_site_parameters}
\title{EMS Site Paramters.}
\format{A data frame with 907694 rows and 22 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
ems_site_parameters
}
\description{
EMS lookup table for which parameters each site has data for.
}
\keyword{datasets}
