% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_check.R
\name{mod_data_check_ui}
\alias{mod_data_check_ui}
\alias{mod_data_check_server}
\title{mod_data_check_ui and mod_data_check_server}
\usage{
mod_data_check_ui(id)

mod_data_check_server(input, output, session, dataset, check, which)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
