% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preview.R
\name{mod_data_preview_ui}
\alias{mod_data_preview_ui}
\alias{mod_data_preview_server}
\title{mod_data_preview_ui and mod_data_preview_server}
\usage{
mod_data_preview_ui(id)

mod_data_preview_server(input, output, session, run_mode, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
