% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{mod_data_upload_ui}
\alias{mod_data_upload_ui}
\alias{mod_data_upload_server}
\title{mod_data_upload_ui and mod_data_upload_server}
\usage{
mod_data_upload_ui(id)

mod_data_upload_server(input, output, session, run_mode)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
