% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ems.R
\name{mod_ems_ui}
\alias{mod_ems_ui}
\alias{mod_ems_server}
\title{mod_ems_ui and mod_ems_server}
\usage{
mod_ems_ui(id)

mod_ems_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
