% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_guideline.R
\name{mod_guideline_ui}
\alias{mod_guideline_ui}
\alias{mod_guideline_server}
\title{mod_guideline_ui and mod_guideline_server}
\usage{
mod_guideline_ui(id)

mod_guideline_server(input, output, session, results)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
