% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset.R
\name{mod_dataset_ui}
\alias{mod_dataset_ui}
\alias{mod_dataset_server}
\title{mod_launch_ui and mod_launch_server}
\usage{
mod_dataset_ui(id)

mod_dataset_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
