% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outlier.R
\name{mod_outlier_ui}
\alias{mod_outlier_ui}
\alias{mod_outlier_server}
\title{mod_outlier_ui and mod_outlier_server}
\usage{
mod_outlier_ui(id)

mod_outlier_server(input, output, session, params, stand_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
