% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rcode.R
\name{mod_rcode_ui}
\alias{mod_rcode_ui}
\alias{mod_rcode_server}
\title{mod_rcode_ui and mod_rcode_server}
\usage{
mod_rcode_ui(id)

mod_rcode_server(input, output, session, data, tidy, clean, outlier, results)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
