% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reference.R
\name{mod_reference_ui}
\alias{mod_reference_ui}
\alias{mod_reference_server}
\title{mod_reference_ui and mod_reference_server}
\usage{
mod_reference_ui(id)

mod_reference_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
