% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_refine_sidebar.R
\name{mod_refine_sidebar_ui}
\alias{mod_refine_sidebar_ui}
\alias{mod_refine_sidebar_server}
\title{mod_refine_sidebar_ui and mod_refine_sidebar_server}
\usage{
mod_refine_sidebar_ui(id)

mod_refine_sidebar_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
