% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results.R
\name{mod_results_ui}
\alias{mod_results_ui}
\alias{mod_results_server}
\title{mod_results_ui and mod_results_server}
\usage{
mod_results_ui(id)

mod_results_server(input, output, session, data, tidy, clean, outlier)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
