% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_site.R
\name{mod_site_ui}
\alias{mod_site_ui}
\alias{mod_site_server}
\title{mod_site_ui and mod_site_server}
\usage{
mod_site_ui(id)

mod_site_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
