% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tidy.R
\name{mod_tidy_ui}
\alias{mod_tidy_ui}
\alias{mod_tidy_server}
\title{mod_standardise_ui and mod_standardise_server}
\usage{
mod_tidy_ui(id)

mod_tidy_server(input, output, session, raw)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
