% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ems_app}
\alias{run_ems_app}
\title{Run the Shiny Application}
\usage{
run_ems_app(dataset = "2yr")
}
\arguments{
\item{dataset}{A string indicating which dataset the app should use.
Possible choices are:
"demo" (loads a small demo dataset that does not require any downloading);
"2yr" (loads and prompts download/update of recent 2 years of data (from 2018-01-01 onwards));
"4yr" (loads and prompts download/update of recent 4 years of EMS data (from 2016-01-01 onwards));
"historic" (loads and prompts download/update of historic EMS data from 1964-01-01 to 2018-01-02);
"all" (loads both recent and historic data);
"upload" (allows user to upload their own dataset in either tidied or raw format).}
}
\description{
Run the Shiny Application
}
