% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_demo_data}
\alias{ems_demo_data}
\title{EMS demo data.}
\format{
A data frame with 250241 rows and 22 variables:
}
\usage{
ems_demo_data
}
\description{
EMS data for three paramters ("Temperature", "pH", "Turbidity") and 1000 sites.
}
\keyword{datasets}
