% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clean.R
\name{mod_clean_ui}
\alias{mod_clean_ui}
\alias{mod_clean_server}
\title{mod_refine_ui and mod_refine_server}
\usage{
mod_clean_ui(id)

mod_clean_server(input, output, session, stand)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
