% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_standardise.R
\name{mod_standardise_ui}
\alias{mod_standardise_ui}
\alias{mod_standardise_server}
\title{mod_standardise_ui and mod_standardise_server}
\usage{
mod_standardise_ui(id)

mod_standardise_server(input, output, session, tidy)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
